+++ packages/assets-controllers/src/token-prices-service/codefi-v2.ts
@@ -228,8 +228,6 @@ export const ZERO_ADDRESS: Hex =
  */
 const chainIdToNativeTokenAddress: Record<Hex, Hex> = {
   '0x89': '0x0000000000000000000000000000000000001010', // Polygon
-  '0x64': '0xe91d153e0b41518a2ce8dd3d7944fa863463a97d', // Gnosis
-  '0x3dc': '0x779ded0c9e1022225f8e0630b35a9b54be713736', // Stable - Native symbol: USDT0
   '0x440': '0xdeaddeaddeaddeaddeaddeaddeaddeaddead0000', // Metis Andromeda
   '0x1388': '0xdeaddeaddeaddeaddeaddeaddeaddeaddead0000', // Mantle
 };
@@ -261,7 +259,7 @@ export const SPOT_PRICES_SUPPORT_INFO = {
   '0x39': 'eip155:57/slip44:57', // Syscoin Mainnet - Native symbol: SYS
   '0x52': 'eip155:82/slip44:18000', // Meter Mainnet - Native symbol: MTR
   '0x58': 'eip155:88/slip44:889', // TomoChain - Native symbol: TOMO
-  '0x64': 'eip155:100/slip44:700', // Gnosis (formerly xDAI Chain) - Native symbol: xDAI
+  '0x64': 'eip155:100/erc20:0x0000000000000000000000000000000000000000', // Gnosis (formerly xDAI Chain) - Native symbol: xDAI
   '0x6a': 'eip155:106/slip44:5655640', // Velas EVM Mainnet - Native symbol: VLX
   '0x7a': 'eip155:122/erc20:0x0000000000000000000000000000000000000000', // Fuse Mainnet - Native symbol: FUSE
   '0x80': 'eip155:128/slip44:1010', // Huobi ECO Chain Mainnet - Native symbol: HT
@@ -278,7 +276,7 @@ export const SPOT_PRICES_SUPPORT_INFO = {
   '0x150': 'eip155:336/slip44:809', // Shiden - Native symbol: SDN
   '0x169': 'eip155:361/slip44:589', // Theta Mainnet - Native symbol: TFUEL
   '0x2eb': 'eip155:747/slip44:539', // Flow evm - Native symbol: Flow
-  '0x3dc': 'eip155:988/erc20:0x779ded0c9e1022225f8e0630b35a9b54be713736', // Stable - Native symbol: USDT0
+  '0x3dc': 'eip155:988/erc20:0x0000000000000000000000000000000000000000', // Stable - Native symbol: USDT0
   '0x3e7': 'eip155:999/slip44:2457', // HyperEVM - Native symbol: HYPE
   '0x440': 'eip155:1088/erc20:0xdeaddeaddeaddeaddeaddeaddeaddeaddead0000', // Metis Andromeda Mainnet (Ethereum L2) - Native symbol: METIS
   '0x44d': 'eip155:1101/slip44:60', // Polygon zkEVM mainnet - Native symbol: ETH